Rename repetitive packages and classes names
