+++ service/switchboardapp/src/test/java/com/switchboard/app/SwitchboardappApplicationTests.java
@@ -1,4 +1,4 @@
-package com.switchboard.switchboardapp;
+package com.switchboard.app;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
